Initial commit
Start of a project which might become an easy way to send a file from
one computer to another on the same network, simply by entering the same
passphrase into both. It is currently named TTT, for "This To There".

So far it's only a diagnostic program to test the "discover the other
computer" functionality.

On computer 1 (-l = listen mode):
./ttt -l -s "my hovercraft is full of eels"

On computer 2 (run in announcer mode, with payload 12345):
./ttt -i 12345 -s "my hovercraft is full of eels"

Computer 2 sends a UDP datagram, encrypted with the passphrase, on every
broadcast-enabled and multicast-enabled interface it can find. Computer
1 receives one of these datagrams, decrypts it, validates it, and finds
12345 in the payload.

This is intended to be a port number on which Computer 1 can then
connect to Computer 2 to start a file transfer. The whole file transfer
part isn't written yet.