tttaccepttest: embryonic version of a connection listener
tttaccept.c contains a very simple tool which sets up an "accept a
connection" context, waits to accept the connection in a non-blocking
way while printing a message to stdout every second, then when it gets
a connection which successfully handshakes*, it prints where that
connection came from and exits.

* There is no actual handshake yet, but there will be when this uses
encryption.

When all this is finished, the side that sends out the UDP announcement
datagrams ("please connect to me, person who also has my passphrase!")
will be simultaneously monitoring a listening socket. The correct
recipient of the announcements will connect in and successfully
handshake using the passphrase, but the announcer also has to deal with
everyone and their dog connecting to the listening socket with
irrelevant or malicious intentions. So any received connection is made
non-blocking and we only service it when it has something to say.

The tttaccept.c code provides that any connection we get which then
just sits there doing nothing won't cause us to block, and any
connection which fails to complete a valid handshake will be silently
closed while we continue to wait for new connections.