tttdiscovertest now establishes session after discovery
tttdiscovertest -l: listen for UDP announcement datagrams, and when we
get one that's valid, connect to the announcement's sender on the
invited port, handshake, and spam a message at them.

tttdiscovertest: send UDP announcement datagrams while listening for TCP
connections on the invitation port number. When we get an incoming TCP
connection which handshakes successfully, receive a message spammed at
us and print it to stdout.

tttdiscovertest -l now also generates a random four-word passphrase if
one is not given on the command line, and tells the user to enter that
passphrase on the other side. tttdiscovertest with no -l now prompts for
a passphrase if one is not given.

"handshake" is still the null handshake whereby a function just returns
success without doing anything else.