Got SSL working - TCP connection now authenticated and encrypted
The TCP connection opened by tttaccept.c and maintained as a session by
tttsession.c is now, by default, encrypted and authenticated with the
pre-shared key. When we initiate a session using ttt_session_connect(),
that's encrypted in the same way.

Recall that this TCP connection is the one initiated by the side that
received a UDP datagram which it decrypted (using the key) to a valid
message containing a port number. It TCP-connects to the sender of that
datagram on this "invitation" port number, and both sides satisfy
themselves they're talking to the right endpoint by doing an SSL
handshake using a pre-shared key derived from the passphrase.

In tttdiscovertest, the side that initiated the TCP connection then
sends a short message down it and exits. The other side receives that
message, displays it and exits.

Next step: instead of sending/receiving a short message, negotiate the
transfer of a file or set of files.