Remove potentially ambiguous words from word list
Passphrases are generated from the list of words in wordlist.txt. This
list has already been the subject of some curation in a past project,
but there are a few words in it that could be ambiguous so we shouldn't
use them.

It's plausible that a ttt-generated passphrase might be communicated
verbally to the user of the other computer, so all the words have to be
unambiguous when pronounced, and not lead to confusion in any other way.

Words which I noticed as having plausible homophones are out, such as
ACCEDE (exceed) and BERRY (bury).

Words which have more than one spelling in common usage are out, such as
COLOUR, LABOUR, REALISE, THEATRE, etc.

Words which might be written with an accented character (CLICHE, TOUPEE)
are out.

Most words with an unstressed -MAN suffix are out because it could be
interpreted as -MEN.

Words which might be misinterpreted as numerals, symbols, or
instructions related to other words, such as TWELVE, TWENTY, PLUS,
MINUS, BRACKET, DASH, CAPITAL, PLURAL etc, are out.

Any words from the NATO phonetic alphabet (DELTA, NOVEMBER, etc), are
out because they might be misinterpreted as a single letter.

Finally, I also removed any words I didn't like the look of or which are
commonly misspelled, such as ACQUIESCE, NECESSARY, RECEIVE etc.

Some suitable three-letter words have been added to the word list. For
some reason this word list only contained words of four or more letters.