Improved error handling
If the sender fails to open a file in order to read and send it, this is
no longer a fatal error. The TTT_MSG_FILE_DATA_END message now has a
body whose format is the same as TTT_MSG_ERROR - it contains a code and
a message string. Normally the code is 0 and the string is empty, but if
we fail to open or read from the file we've said we're going to send, we
send a TTT_MSG_FILE_DATA_END with an appropriate error.

The receiver counts the number of files which were omitted in this way,
and if this is nonzero, writes a summary at the end and exits with
status 1.

If on the receiving side we fail to write out any file (fopen() or
fwrite() returns an error), this is still a fatal error. The most likely
causes of this I can think of are that we don't have permission to write
to the destination directory, or the disk is full. In either case,
there's no point in continuing with the rest of the files.

These can be tested by setting the ttt_random_file_*_failures global
variables to 1.

Reword protocol.h to say that if either side receives an unrecognised
tag, or a tag which isn't valid at that time, this is always a fatal
error and the session must be aborted. It doesn't work to just reply
with an error, because this unknown tag might be part of a longer
yet-to-be-defined message sequence and the sender isn't expecting
multiple error replies. If we want to allow forward compatibility with
yet-to-be-defined messages, we should introduce a capability
negotiation stage at the start of the session where each side tells the
other what it supports.