File transfer context structure for options, progress callback etc
struct ttt_file_transfer is a context structure which tttpush and
tttpull initialise and pass to ttt_file_transfer_session(). There are
two initialisation functions: one for the side that starts as sender,
the other for the side that starts as receiver.

After initialisation, the context structure can be set up with various
options and callbacks, detailed in filetransfer.h. This saves us from
having to pass an ever-growing list of parameters to
ttt_file_transfer_session().

Progress reports are now done by a callback in this context structure
so the caller can supply their own progress output code if necessary.

ttt pull: --confirm option will pause after receiving the metadata set,
show the first ten file names and the total file count and size (if
given), and give the user the option to accept or reject them.

ttt push: --send-full-metadata affects what we send in the metadata
message sequence, before we start sending any data. If
--send-full-metadata is given, we send metadata for every file in the
list. Otherwise, we send only the total file count and size. Really
this is only useful if --confirm was supplied on the pulling end.

ttt push and pull now show the peer's IP address when a session is
established.