Adapt ttt to work on Windows
Introducing a load of #ifdef WINDOWS ... #else ... #endif blocks all
over the place.

compile_win.sh, which for some reason isn't a Makefile, uses mingw to
compile a Windows binary named ttt.exe. This needs an OpenSSL
installation compiled with mingw, from which it plucks the libssl.a and
libcrypto.a files and compiles them statically into the finished binary.

Magic command to cross-compile OpenSSL on Linux for Windows with mingw,
having downloaded and extracted OpenSSL and entered its top-level
directory:

    ./Configure --cross-compile-prefix=x86_64-w64-mingw32- mingw64
    make

That was easier than I expected.

Code changes:

netif.c is almost completely rewritten for Windows because the existing
interface-discovery code is heavily Unix-specific.

ttt_socket_error() is like ttt_error() except that it doesn't take an
errno argument, it just assumes you want to use errno and does so. On
Windows it uses WSAGetLastError() instead of errno.

Interface detection currently only lists IPv4 addresses, and when we
listen on a TCP socket, this is currently only on the AF_INET (IPv4)
address family. This was to fix a problem where the Windows side was
only announcing on IPv4, so the Linux machine tried to connect to the
Windows machine's IPv4 address, but the Windows machine was only
listening on an IPv6 address. We'll probably want to listen on both
before re-enabling IPv6.

ttt_make_socket_non_blocking() and ttt_make_socket_blocking() are now
defined in utils.c for both Windows and Linux, so we don't have fcntl()
calls littered everywhere.

On Windows, the interface discovery change can give us a list containing
irrelevant interfaces we don't want to use for sending announcement
datagrams, such as link-local or localhost interfaces. If we try to
enable multicast or send datagrams on these it failed and we bailed out,
even though there might have been at least one useful interface in the
list. So now, when preparing our list of sockets, if setting the
IP_MULTICAST_IF option fails, simply close the socket and set
ctx->sockets[...] = -1. If all the sockets fail, we fail, but as long
as we have at least one we continue. Everywhere in discover.c that
iterates over that list of sockets now knows to skip over any entries
that are -1.