More Windows fixes, this time when pushing files
When listening for announcement datagrams, we call setsockopt() on our
socket with IP_ADD_MEMBERSHIP to get it to subscribe to the right
multicast address on every interface. However, some of these interfaces
will fail to have anything to do with multicast. Soldier on unless they
all fail.

Also, make sure when we get a socket it's an IPv4 socket, until we've
got IPv6 multicast sorted out.