Improved Windows socket error handling and discovery callbacks
ttt_socket_error() now prints the string associated with a Windows
socket error, rather than a cryptic five-digit number.

ttt_discover_and_connect() now takes some callback arguments so that the
caller can be notified when:
    * We successfully set up the socket for receiving announcements.
    * We receive any announcement datagram, valid or otherwise.

tttpush uses these to tell the user what passphrase we generated only
after we've set up our announcement-receiving socket, and to report to
the user that we've received a valid announcement and found our peer,
and (if verbose) to report that we've received an invalid announcement.