ttt_chmod() function to cope with Windows or Linux chmod()
The Windows _chmod() function behaves slightly differently from Linux's
chmod(), mainly in that it only has "can you read it" and "can you write
it" bits rather than the Unix -rwxrwxrwx thing. Make a function
ttt_chmod() which we can call on both platforms, which calls the
appropriate function with the right arguments.