IPv6 support for discovery and connection
Both the pushing and pulling side of ttt now support IPv6. By default,
the pulling side sends broadcast and multicast announcements on all
usable IPv4 and IPv6 interfaces, and the pushing side joins both an
IPv4 and IPv6 multicast group to receive announcements.

Then when the pushing side receives a valid announcement, it connects to
the pulling side on the source address of the valid announcement,
whether that be IPv4 or IPv6.

New options to ttt push:
    --no-ipv4, --no-ipv6: disable IPv4 or IPv6.

New options to ttt pull:
    --no-ipv4, --no-ipv6: disable IPv4 or IPv6.
    --no-broadcast, --no-multicast:
        disable broadcast or multicast announcements.

Still to do:

Allow the user to change the multicast destination/listen address by a
command line option.

Weird IPv6 issue when pushing on Windows and pulling on Linux, with
--no-ipv4 on the Linux end. Windows struggles to connect to Linux on an
IPv6 address. The connection attempt hangs and times out. However, if
while this connection attempt hang is going on, you ping the Linux box
on an IPv6 address in another window, the connection magically works.