Make connection attempts non-blocking and remove duplicate addresses
Currently, when we're pushing a file, we wait to receive announcements
from the pulling side, then we connect to the source address of that
announcement, blocking until that succeeds or fails.

The problem with this is that the pulling side could have sent an
announcement on many different interfaces. If the first announcement we
receive is on an interface we can't connect to them on for whatever
reason, the connect() call may hang even though we could be receiving
other announcements from addresses we could successfully connect to.

So now, in ttt_discover_and_connect(), when we receive an announcement
datagram, we start a non-blocking connect attempt and keep that in a
list (connect.c). We then go back to listening for further
announcements alongside listening for events on that list of partially
connected sockets. If we receive any other announcements we set up more
non-blocking connect attempts and add them to the list.

Eventually, when one of the connection attempts succeeds, the select()
call in ttt_discover_and_connect() will notice, we'll call
tttmcctx_run() and it will return the newly-connected session.

When this happens, we do the handshake with that connection (blocking
on its completion). If the handshake succeeded then we take that
session and close all the other partially-set-up sessions, and if it
failed then we forget it and continue listening for
announcements/connection completions.

We could consider doing the handshake in a non-blocking way as well, the
same as on the accepting side (see accept.c) but then we'd have to solve
the problem of both sides getting a different idea of which session's
handshake completed first.

Additionally, in netif.c:
Remove duplicate addresses from the list of network interface addresses,
should they exist.

Also, announcements we send on the various interfaces must come from the
relevant interface address, not any old address. An interface can have
many addresses, and if we send one announcement for each of these
addresses, the source address must be correct so the receiver knows
where to connect to if it receives that announcement. So in
tttdactx_init(), when we associate a socket with each interface, we
bind() it to the relevant address.