Skip directory if first readdir() fails
Some directories on Windows appear to be phantom hidden directories
which cannot be read. For example, the "My Music" directory in a
user's Documents folder. These directories can be opened with
opendir() but the first readdir() fails. This is treated as a fatal
error.

If the first readdir() on a directory fails, assume it's one of these
directories and skip it at no penalty.