Support for FIFO files and directory metadata
Directories are now sent as file entries, after all the files in that
directory have been sent. The metadata for the directory contains the
modified time and the permissions of the directory, so the receiver
can apply these to the directory. This also means that empty directories
in a directory structure now get created on the receiver.

FIFO files are also supported (Linux only).

On Windows, when we receive files we now do some rudimentary character
replacement on filenames so they don't contain any characters which are
illegal in Windows filenames.