If we push a directory name, send the directory not just its contents
If we push /foo/bar/baz, then on the other side when we run "ttt pull"
it's less surprising for the directory "baz" to be dropped into the
current directory, rather than smearing the contents of "baz" all over
the current directory and leaving the pulling user to clean up the mess.

Changes to ttt_dir_walk() to get rid of the "initial_prefix" thing and
replace it with an offset into the pathname. In ttt_dir_walk_aux() this
can now be used as an offset into the current pathname to locate the
start of the last directory component of the original pathname supplied
to the top-level ttt_dir_walk() call.

So if we've pushed /foo/bar/baz, and deep in a recursive ttt_dir_walk()
call the current path is /foo/bar/baz/wibble/fish/balloons.txt, the
offset points to the start of "baz". This is also used to create the
pathname we send to the pulling side, so they see the file name as
"baz/wibble/fish/balloons.txt".