Default listen port is now same as default discovery port
The default TCP port the pulling side listens on and includes in the
payload of the announcement packet is now the same as the UDP port
we send the announcements on. ttt pull can still change this by
specifying --listen-port. To get the old behaviour of just choosing any
free TCP port, use --listen-port 0.

Don't call the multicast group address a "rendezvous" address, to avoid
confusion with the specific meaning of the word "rendezvous" in the
context of IPv6 multicast.

Invert the sort order of the list of network interface addresses
returned by netif.c, so we announce to IPv6 addresses first, and the
higher-numbered addresses first. This has the effect that we announce
from the link-local IPv6 addresses first (fe80::/10), then from the
unique local addresses (fc00::/7) then the unique global addresses.

Talking of global addresses: we now don't announce to these, or enable
the associated interfaces to receive multicast packets, unless
--include-global is given. By default we announce to private IP ranges
only (10.x.x.x, 172.(16-31).x.x, 192.168.x.x, link-local 169.254.x.x,
and their IPv6 equivalents fc00::/7 and fe80::/10).