Options to prompt for passphrase, and don't hide it by default
Not echoing the passphrase is getting annoying. Turn that off by default
and allow the user to re-enable it with --hide-passphrase.

"ttt push --prompt-passphrase" is introduced to tell ttt push to prompt
for a passphrase (asking for it twice if --hide-passphrase is given)
rather than generating one from the word list.