Add unit tests using the CUnit framework
filetransfer.c and netif.c now have unit tests for functions which make
sense to be unit-tested. These unit tests require libcunit.

To run the unit tests:

1. Install libcunit1 and libcunit1-dev if they're not already installed.

2. Compile with -lcunit and -DTTT_UNIT_TESTS=1. This can be done by
   setting CUNIT=1 at the top of the Makefile and running "make ttt"
   as usual.

3. ./ttt test

The compile-for-Windows script requires a path to a mingw-built version
of libcunit. I built this by downloading the CUnit source, and
following (mostly) the instructions, except I also ran "libtoolize" and
"automake --add-missing", before passing the argument
--host=x86_64-w64-mingw32 to ./configure, then building libcunit
with "make".