Callback function to report sent announcement
The caller of ttt_discover_and_accept() can now have a callback function
called every time we send an announcement packet. The callback is
defined in discover.h as tttda_sent_announcement_cb, and takes as its
parameters a cookie pointer, the number of announcement rounds we've had
so far, the announcement sequence number within this round (in each
round we send an announcement from every network interface IP address),
and the source and destination addresses of the announcement packet.

tttpull.c uses this to warn the user that a number of announcements have
been unanswered and that they might want to check they got the
passphrase correct.