Optional timeout for ttt push
ttt push now takes a --timeout option (-t for short) which takes a
number of seconds as its argument. If it takes longer than this to
receive a valid announcement and complete a connection to the other
host, we fail.

ttt pull --announcement-interval now takes a number of seconds rather
than milliseconds. As with ttt push --timeout, this does not have to
be an integer. It is parsed as a double.

timeval_add and timeval_diff are now used in more than one place so
they're moved to utils.c with suitable unit tests for them.