Use announcement packet's salt in derivation of session key
Currently, the UDP announcement packet sent from the pulling side to the
pushing side contains an eight-byte salt. This is to be combined with
the shared secret passphrase to derive the key to encrypt/decrypt the
announcement packet.

However, the key we pass to OpenSSL during the TLS handshake, which
authenticates and encrypts our session, has no salt on it. It's only
derived from the passphrase.

For extra security to defeat a hypothetical rainbow table of all the
keys from all possible generated passphrases, let's derive the session
key in the same way as we derive the announcement packet's encryption
key, using the same salt.

All announcement packets from the same announcement context (tttdactx)
must have the same salt and therefore the same encryption key, because
when we receive a connection in to our listening TCP socket, we don't
know which announcement packet the other side is responding to.

The session key, derived from ttt_passphrase_to_key(), must now be
passed to ttt_session_init() when creating the session. The pushing side
knows the salt because it's in the announcement packet it received, and
the pulling side knows the salt because it sent the announcement packet.

session.c now no longer has one global session key which applies to all
sessions. We only had this because OpenSSL's callbacks for supplying a
pre-shared key didn't have a so-called "cookie" parameter we could use
to tell the callback which session context this was for. However, they
do have the SSL pointer (SSL *ssl), so we keep a mapping, in the form of
a linked list, of all the live SSL objects we have and the session key
associated with them.

ttt_session_tls_init() adds the key to this list. The callbacks call
ttt_find_session_key() to look up the session key given the SSL *ssl
parameter, and ttt_remove_session_key() is called when a session is
destroyed.

We're now a step closer to the code being usable for multiple TTT
sessions within the same process, but note that the functions
ttt_set_session_key(), ttt_find_session_key() and
ttt_destroy_session_key() are not (yet) thread-safe. If called
concurrently they can and will interfere with each other.

Our key length is now #defined in utils.h as TTT_KEY_SIZE, and is
currently 32. This is a slight improvement on having the magic number 32
all over the place.

Remove ttt_session_connect() - nothing calls it.