ttt push can read from stdin, ttt pull can write to stdout
ttt push now takes "-" as a filename argument which causes it to read
from stdin and send that.

ttt pull now takes an option --output-file which, if supplied, causes
all received files to be written to the named file, one after the
other. If the argument to --output-file is "-", we write to stdout.
If you give "-o -" to ttt pull, this is taken as "--output-file -".
You can't supply both -o and --output-file. You can output files to a
directory, or all the data to one file, but not both.

If "ttt push" sends data from stdin, but "ttt pull" is outputting to a
directory as normal, the stdin file appears as a single regular file
named "stdin" in the top-level output directory, with a modified-time
of now and a mode of 0644. The size of stdin is given in the
TTT_MSG_FILE_METADATA message is -1 (not known).

If "ttt pull" is writing everything to a file or stdout (--output-file
given), and "ttt push" sends multiple files, all pushed files get
written to that one destination file, and separating them out into
their original files if required is the user's problem.

The progress indicator will not show a percentage or total size if one
of the files in the set is stdin, because its size is unknown before
we've received the whole file.

Other stuff:

Both ttt push and ttt pull now support -q or --quiet to disable the
progress indicator and other non-essential output.

ttt_mkdir_parents() now checks if the directory or parent already
exists, and does nothing if so. This simplifies ttt_receive_file_set()
a bit.