Handle Unicode filenames on Windows
On Windows, file names are Unicode strings. The library calls such as
fopen(), stat() expect the filename to be in whatever codepage Windows
has decided it prefers, but these functions also have a w-variant (e.g.
_wfopen()) which take the filename as a wchar_t * rather than a char *.

On Linux, file names are an arbitrary array of bytes. This leads to a
"good luck with that" approach to parsing them as characters without
knowledge of the character encoding used by whatever created them.

This poses an obvious problem when we want to send files from one of
these operating systems to another.

The rule I've decided on for the wire protocol, when the filename is
sent from the sender to he receiver in the TTT_MSG_FILE_METADATA
message, is that the filename must be encoded as UTF-8. Converting that
name to whatever character encoding is suitable for the receiving OS is
the receiver's problem.

On Linux:
    * When sending a file, we assume the filename was probably intended
      to be read as UTF-8. If it isn't valid UTF-8, we replace any
      troublesome bytes with '_'. Now it is valid UTF-8! Easy!
    * When receiving a file, we take the UTF-8 name in the metadata
      message and simply pass that as-is to fopen() etc.

On Windows:
    * When sending a file, we have the filename as a wide-character
      string, and so we convert that string to UTF-8 for sending.
    * When receiving a file, we take the UTF-8 filename we're given
      and convert that to a wide-character string. We use that
      wide-character string to refer to the local file we create when
      we call the W-equivalents of fopen(), stat() etc.
    * On Windows, wide-character strings are UTF-16, and wchar_t is
      two bytes.

To facilitate all this, localfs.c and localfs.h are introduced.
This defines TTT_LF_CHAR to mean the character type used to build local
filenames as we want to represent them on the host OS. It's char on
Linux and wchar_t on Windows.

localfs.c contains a set of wrapper functions ttt_fopen(), ttt_stat()
etc, which expect a TTT_LF_CHAR * where a filename is required, and they
call the appropriate Windows or Linux library function. We also have
string handling functions like ttt_lf_copy() (strcpy or wcscpy).

Functions we already had in order to abstract away Linux/Windows
differences like ttt_chmod() and ttt_mkdir_parents() are moved from
utils.c to localfs.c.

localfs.c also provides some helpful functions for converting from a
TTT_LF_CHAR-string to UTF-8 or vice versa.

There are a few other symbols defined in localfs.h, such as
TTT_LF_PRINTF, which is defined as the printf format specifier to use
with a TTT_LF_CHAR * ("%s" for char *, or "%S" for wchar_t *).

Listing directory entries on Windows is now done with FindFirstFileW()
and FindNextFileW(), but we wrap that in functions called ttt_opendir(),
ttt_readdir(), and ttt_closedir(), which behave more like the Linux
functions opendir(), readdir(), and closedir().