Introduce two-way hello exchange before TLS handshake
A bit of forward compatibility here.

Before the TLS handshake, the client and server now both send to each
other a 12-byte hello message (client sends first), which contains a
magic number, the minimum supported protocol version, the maximum
supported protocol version, and a 32-bit "flags" word.

Currently the only protocol version in existence is version 1 and the
flags word has no meaning yet so it's zero.

However, if in the future we change the protocol in a potentially
incompatible way, we can either use one of the flags bits to indicate
support for an optional extension, or bump the protocol version number.

If the minimum supported protocol version of one side is greater than
then maximum supported protocol version of the other side, there is a
protocol version mismatch and both sides should close the session and
return an error. Otherwise, the session should use the lower of the
server's maximum supported version and the client's maximum supported
version.