Use TTT_LF_PRINTF not "%s" when printf-ing a TTT_LF_CHAR *
When we supply a local filename (TTT_LF_CHAR *) to a printf-like
function, we must use TTT_LF_PRINTF as the format specifier, not "%s".
"%s" only works if local filenames are multibyte strings, and on Windows
they're Unicode strings. TTT_LF_PRINTF is "%ls" or "%s" as appropriate.

Fix the eleventy billion places we get this wrong.