Send salt for session key in client hello message
Both sides combine the shared passphrase with a random but cleartext
salt to derive the pre-shared key. Since commit 51397663 this salt is
taken from the salt in the UDP announcement packet sent by the puller
(TCP server) and received by the pusher (TCP client).

Now that we have a hello exchange stage before the TLS handshake, it
makes sense for the client to generate and send the random salt in its
hello message, rather than requiring all a puller's announcement packets
to have the same salt. So undo most of what we did in commit 51397663.

All announcement packets can now have different salts and the salt no
longer needs to be preserved and passed down to the connection-accepting
context, session object etc. After the hello exchange, both sides have
the salt and so both can combine it with the passphrase to derive the
key they'll use in the TLS handshake.