Fix passphrases again, and allocate struct ton_msg on heap not stack
A commit earlier today forgot to copy passphrase_length along with
passphrase in session.c and accept.c, with the result that it was
using a zero-length passphrase on both sides. Fixed and put a warning in
so it tells us if it's using a zero-length passphrase in future.

filetransfer.c now allocates struct ton_msg objects on the heap rather
than on the stack. Seems more suitable for something that's 64kB.