ton_mkdir_parents() was trying to create "C:" on Windows
When we pass "c:\users\graeme\long\path" to ton_mkdir_parents(), that
function tries to create every directory along the path if it doesn't
already exist.

The problem is that it thinks the first component is "c:", and when
ton_stat() fails to stat that, it assumes it doesn't exist and tries to
create it. It is no surprise to learn that this fails.

Don't try to create a directory component on Windows if its last
character is ':'. This character isn't allowed in a normal filename on
Windows, so if we see it we can be pretty confident it's part of a drive
letter and we should just move on to the next directory component.