Add --confirm to man page, put longopts array in sensible order
"ton pull --confirm" is something I forgot I had implemented, and gives
the user the opportunity to cancel the file transfer once they see the
number and size of the files about to be sent.

Add --confirm to the ton pull man page, and remove the bit of the ton
push man page which says that functionality doesn't yet exist.