Add -f/--force option, otherwise ask to overwrite existing files
filetransfer.c can now be given a callback function to call whenever
we're about to receive a file. It is given details of the file, and
can decide whether to accept the file, overwriting any existing file
with the same name (TON_FT_ACCEPT), receive the file but skip writing
it to disk (TON_FT_SKIP) or abort the whole transfer (TON_FT_ABORT).

ton pull uses this to ask the user whether they want to overwrite an
existing file with one we're about to get. Use "ton pull -f" if you want
the old behaviour, which overwrites any existing files without
prompting.

The prompt has five choices: overwrite the file, keep the existing file,
overwrite this and any other file, keep this and all other files, or
quit the transfer.