Introduce "ton version" showing build timestamp and git commit hash
The Makefile and the compile_win.sh script now shoehorn the stringified
output of "git rev-parse HEAD" into ton as TON_GIT_COMMIT_HASH. This
value, along with the build date and time, can be printed with
"ton version".

This is to help resolve any confusion about what version of ton
someone has.

Example output:

    graeme@panini:~/code/ton$ ton version
    ton
    git commit hash: 35ca28755f8d00b473bdbc68d9fb4cfde20da6f0
    This binary was compiled on Sep 27 2022 at 20:58:58

Later on, an actual version number could be added to that first line of
output.