Generalise localtime_r() and don't set metadata on a skipped file
ton_localtime_r() is now defined, which calls localtime_r() if that's
available and localtime_s() on Windows. It behaves like localtime_r()
except it returns an integer success/failure value rather than a
pointer.

Previously, if we skipped a file because a file with that name already
existed and the user wanted to keep the existing one, we incorrectly set
the file's mode and modified time to that of the transferred file. We
now only set the file's mode and modified time if we actually saved the
file.