End-to-end test scripts
To run the tests, run testpush.py in one terminal and testpull.py in the
other. If there are any problems, one or both sides will throw an
exception. If the other side didn't throw an exception then its next
test will time out after 30 seconds.

The tests themselves are defined in testcommon.py. Each test has
command line arguments to be used on the pushing side and pulling side,
and a definition of the directory structure the pushing side should
create and send. The pulling side checks the directory structure and
file contents are correct.

It's a bit rough and ready but it's automated testing, which we didn't
have before.