Build wordlist.txt into the executable rather than autogenerating C
wordlist.txt is the list of words we choose from when generating a
random passphrase. Until now the Makefile rules caused it to be fed
through a Python script which generated a C file containing a massive
array of string literals.

wordlist.txt is now linked directly into the executable file and
accessed from src/wordlist.c. The first time we look up a word, we
initialise the list by making a copy of the file's contents in which we
lowercase all the letters and replace newlines with '\0'. This gives us
a series of null-terminated strings. We then build an array of pointers
to these strings and then we have everything we need to return the word
with a given index.

The functions declared in wordlist.h can be used by the rest of the code
as before - their behaviour hasn't changed.

The result of this is that Python is no longer required to build ton.