Add support for transferring symlinks
Symlinks are now transferred by ton and recreated on the receiving
side with the same target path.

A file's status as a symlink is indicated in the TON_MSG_FILE_METADATA
protocol message by the file mode field as returned by lstat()
(S_ISLNK(mode)).

Protocol change:
The TON_MSG_FILE_METADATA now has a new string field on the end which
contains the symlink target path. If the file isn't a symlink then this
can be blank and it will be ignored. This field is optional - if the
message ends before the field then we take the symlink target path as an
empty string. This means we don't need to bump the protocol version
number and the wire protocol is still compatible with old versions of
ton which don't send this extra field in the TON_MSG_FILE_METADATA
message.

Note, however, that old versions of ton dereference symlinks and send
the actual file, so old versions of ton would never send a symlink
anyway.

Symlinks do not apply to Windows. No attempt is made to convert them to
Windows shortcuts. If ton on Windows receives a file describing itself
as a symlink, it will be ignored just like any other kind of file we
don't support, and not created on the receiving system.

The symlink's target path is copied verbatim from the sending system,
regardless of whether that is a relative or absolute path. This is no
problem if the symlink contains a relative path referring to a file
which is being included in the transfer, but if it's an absolute path
which doesn't happen to exist on the receiving system then you'll end
up with a dangling symlink. This is a tricky problem to fix for all
cases, so for now I've decided this is the user's problem.