Don't follow symlinks for files named on command line
Don't use realpath() to resolve pathnames given to us, because that also
resolves symlinks and we don't want it to do that. If a user tries to
ton push something with a wildcard in it and a symlink is one of the
files matched by the wildcard, we want to send the symlink not the file
to which it points.

This has the possibly unwanted side-effect that if you do "ton push foo"
and foo happens to be a symlink, you now push the symlink and nothing
else. If it becomes a problem we could add an option to say "resolve
symlinks for files named on the command line", but for now it doesn't do
that.