Resolve . and .. components of ton-pushed paths, but not symlinks
Now I remember why ton push called realpath() on the paths named on the
command line. It's so that if we're in the directory "/foo/bar" and we
run "ton push ." we transfer a directory called "bar" and not a
directory called ".".

This desirable behaviour was broken by commit 1c3e1dc9, "Don't follow
symlinks for files named on command line".

However, we still don't want to follow symlinks in paths given to ton
push. So rather than using realpath(), we now use ton_dedotify_path(),
which canonicalizes the path (makes the path absolute, replaces
repeated directory separators with a single separator, and resolves
"." and ".." components) but doesn't follow symlinks.

Unit test added for ton_dedotify_path().